修改节点设置；修改表结构；完成仓库相关数据设置；设计商品sku

+++ controllers/base.go
@@ -62,7 +62,7 @@ func (bc BaseController) CurrentUser() *proto.User {
 // @param action is action summary
 // @param details are more addtional detail
 func (bc BaseController) OperationLog(action string, details ...string) {
-	uid := 0
+	uid := int64(0)
 	nickname := ""
 	user := bc.GetSession(cSessionUserInfoKey)
 	if user != nil {
